use cdecl instead regparm=3

+++ string.h
@@ -1,8 +1,8 @@
 #ifndef __STRING_H__
 #define __STRING_H__
 
-void *memset(void *buf, int value, unsigned size);
+void *memset(void *buf, int value, unsigned long size);
 
-void *memcpy(void *dst, const void *src, unsigned size);
+void *memcpy(void *dst, const void *src, unsigned long size);
 
 #endif /*__STRING_H__*/
